backend/requirements.txt
# Core dependencies
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18

# AI/LLM clients
openai>=2.16.0,<3.0.0
anthropic>=0.77.0,<1.0.0
dspy-ai>=2.5.0

# Web scraping and parsing
httpx>=0.27.0
beautifulsoup4>=4.12.2
pypdf2>=3.0.1
python-docx>=1.1.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
aiosqlite>=0.20.0  # For local SQLite development
alembic>=1.14.0
supabase>=2.3.0

# Background tasks
celery[redis]>=5.4.0,<6.0.0
redis>=5.2.0

# Authentication
fastapi-clerk-auth>=0.0.9

# Observability
sentry-sdk[fastapi]>=2.0.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
opentelemetry-instrumentation-celery>=0.50b0
opentelemetry-exporter-otlp-proto-grpc>=1.29.0

# Analytics
posthog>=3.0.0

# Email
resend>=2.0.0

# WebSocket pub/sub
broadcaster[redis]>=1.0.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
structlog>=23.2.0

# --- Phase 3: Agent framework (ADR-1: Custom chosen over LangGraph) ---
langfuse>=2.0.0
celery-redbeat>=2.3.3
vcrpy>=6.0.0
pytest-recording>=0.13.0

# Testing dependencies
schemathesis>=3.30.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-env>=1.1.1
httpx-mock>=0.3.1
faker>=20.0.3

# Development tools
ruff>=0.9.0
mypy>=1.7.0
pre-commit>=3.5.0
